non-root: add infrastructure to see current UID
The idea behind the rootless CRIU code is, that CRIU does one getuid()
call and adds the value to the global opts structure.

If CRIU is running as non-root this field will only be filled with a
non-zero value if the CRIU binary has the necessary capabilities.

Currently this is either CAP_SYS_ADMIN or CAP_CHECKPOINT_RESTORE.

Signed-off-by: Adrian Reber <areber@redhat.com>