non-root: add function to check for capabilities
This adds the function check_uid() which checks if CRIU is running as
non-root or not. If it is running as root, nothing changes.

If it is running as non-root, CRIU looks if the binary has the necessary
capabilities set to run as non-root. Either CAP_SYS_ADMIN or
CAP_CHECKPOINT_RESTORE.

If one of the capability is set the value of opts.uid is set to the
current UID and this field can then be used in the remaining code to
decide if it is running as non-root or not.

opts.uid will not contain a non-zero value if the necessary capabilities
are not set.

Signed-off-by: Adrian Reber <areber@redhat.com>