travis: enable lazy pages tests again
The lazy pages tests were checking if the kernel is newer than 4.18
using:

[ "$KERN_MAJ" -ge "4" ] && [ "$KERN_MIN" -ge "18" ]

This unfortunately did not work for kernel 5.3 used by Travis right now.

Only checking if the major version is >= 5 should make sure the kernel
has all necessary features.

This means, however, lazy pages tests will not be run on CentOS/RHEL 8,
but as this is about Travis, it should not be a problem.

Signed-off-by: Adrian Reber <areber@redhat.com>