cgroup: fix --manage-cgroups=ignore
Using '--manage-cgroups=ignore' was still trying to restore cgroup
properties at certain places.

Restoring a OCI container in a different cgroup (than the one it was
checkpointed from) resulted in CRIU trying to put the container in the
original cgroup but the container runtime (runc, crun) was putting it
another cgroup. A second checkpoint of the container with two different
cgroup settings did no longer work.

With this change CRIU really completely ignores the cgroup and the
container can be checkpointed and restored multiple times even if the
cgroup changes.

Signed-off-by: Adrian Reber <areber@redhat.com>